Define colvarproxy_replicas in dedicated header
